chore(refactor) - import classes from appropriate modules
