+++ hirola/front/tests/dashboard/test_views.py
@@ -1,6 +1,7 @@
-from front.base_test import *
+from front.base_test import (BaseTestCase, Client, PhoneList, image)
 from front.tests.test_users import UserSignupTestCase
-
+from front.models import (Review, User, ShippingAddress, Order, OrderStatus,
+                          Cart)
 
 class DashboardTemplate(BaseTestCase):
