+++ hirola/front/tests/test_views.py
@@ -1,8 +1,8 @@
-from front.base_test import (BaseTestCase, PhoneList, PhonesColor, image,
-                             Currency, cache)
+from front.base_test import (BaseTestCase, image, Currency, image, cache)
 from front.errors import (category_image_error, phone_category_error)
 from .test_users import UserSignupTestCase
 from .test_cache import phone_form
+from front.models import (PhoneList, PhonesColor, Currency)
 
 
 class LandingPageViewsTestCase(BaseTestCase):
